requirements.txt
py-cord==2.6.1
squarecloud-api==1.0.0
qrcode[pil]==7.4.2
requests==2.31.0
PyYAML==6.0.1
aiofiles==23.2.1
